+++ resources/views/layouts/guest.blade.php
@@ -48,6 +48,8 @@
 
     <link rel="stylesheet" href="{{ asset('vendor/isproject/isproject.css') }}">
 
+    @include('isproject::partials.accent')
+
     @include('isproject::partials.pwa')
 
     @stack('styles')